Add tests for invalid data
Problem: We have 'invalid' data sitting in the repository, but we aren't
testing it. Also, the byte length check is validating the number of
characters instead of the number of bytes in the payload.

Solution: Add tests for the invalid data and fix up the code so that it
passes all (except one) test.