Fix signature regex bug
Problem: The signature regex has ignored backslashes, but this bug is
masked by another bug where a function is returning an error on failure
(when it *should* return something falsey on error), meaning that the
signature checking code doesn't work at all.

Solution: Fix the signature checker regex and the backward return value.