Fix 'invalid HMAC key' error for buffers
Problem: Previous versions of this module allowed HMAC keys as buffers,
but there weren't any tests or documentation which meant that they were
eventually broken.

Solution: Add a test to identify the problem and fix the HMAC key
validation function to ensure that it validates HMAC keys that are
expressed as buffers.

Fixes: https://github.com/ssb-js/ssb-validate/issues/23